chore: reorganize test files into two folders: in_game & out_of_game
