+++ tests/e2e/specs/in_game/basicMoves.spec.js
