+++ tests/e2e/specs/in_game/layout.spec.js
