+++ tests/e2e/specs/in_game/sevens.spec.js
