+++ tests/e2e/specs/out_of_game/authentication.spec.js
@@ -1,4 +1,4 @@
-import {  assertSnackbarError } from '../support/helpers';
+import {  assertSnackbarError } from '../../support/helpers';
 
 const validEmail = 'myCustomEmail@gmail.com';
 const validPassword = 'passwordLongerThanEight';
