+++ tests/e2e/specs/out_of_game/lobby.spec.js
